refactor(gateway): keep cached turn replay with dedupe
