+++ src/gateway/agent-turn/agent-dedupe.ts
@@ -1,6 +1,9 @@
+import { asOptionalRecord } from "@openclaw/normalization-core/record-coerce";
+import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
 import type { GatewayRequestContext } from "../server-methods/types.js";
 import { setGatewayDedupeEntry } from "./agent-job.js";
+import type { AgentTurnIo } from "./types.js";
 
 export function resolveAgentDedupeKeys(params: {
   idempotencyKey: string;
@@ -141,4 +144,44 @@ export function setAbortedAgentDedupeEntries(params: {
       },
     },
   });
+}
+
+export function replayAgentTurnIfCached(params: {
+  preflight: { agentDedupeKeys: readonly string[]; runId: string };
+  context: GatewayRequestContext;
+  io: AgentTurnIo;
+}): boolean {
+  const { agentDedupeKeys, runId } = params.preflight;
+  const cached = readGatewayDedupeEntry({
+    dedupe: params.context.dedupe,
+    keys: agentDedupeKeys,
+  });
+  if (!cached) {
+    return false;
+  }
+  if (cached.ok && isAcceptedAgentDedupePayload(cached.payload)) {
+    const cachedRunId = normalizeOptionalString(cached.payload.runId) ?? runId;
+    const cachedSessionKey = normalizeOptionalString(cached.payload.sessionKey);
+    const cachedAgentId = normalizeOptionalString(cached.payload.agentId);
+    const cachedRuntime = asOptionalRecord(cached.payload.runtime);
+    const admissionPending = typeof cached.payload.reservationId === "string";
+    params.io.emitAcceptance(
+      [
+        true,
+        {
+          runId: cachedRunId,
+          status: "in_flight" as const,
+          ...(cachedSessionKey ? { sessionKey: cachedSessionKey } : {}),
+          ...(cachedAgentId ? { agentId: cachedAgentId } : {}),
+          ...(cachedRuntime ? { runtime: cachedRuntime } : {}),
+          ...(admissionPending ? { admissionPending: true } : {}),
+        },
+        undefined,
+      ],
+      { cached: true, runId: cachedRunId },
+    );
+  } else {
+    params.io.emitAcceptance([cached.ok, cached.payload, cached.error], { cached: true });
+  }
+  return true;
 }
