+++ src/gateway/agent-turn/agent-turn-service.ts
@@ -1,4 +1,3 @@
-import { asOptionalRecord } from "@openclaw/normalization-core/record-coerce";
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import { ErrorCodes, type AgentWaitParams } from "../../../packages/gateway-protocol/src/index.js";
 import { MAIN_SESSION_RECOVERY_WORK_ADMISSION_OWNER } from "../../agents/main-session-recovery/main-session-recovery-admission.js";
@@ -25,7 +24,7 @@ import { prepareSkillLibrarySessionCreation } from "../skill-library-session.js"
 import { createAgentAdmissionController } from "./agent-admission-controller.js";
 import { prepareAgentContentPhase } from "./agent-content-phase.js";
 import { createAgentDedupeLifecycle } from "./agent-dedupe-lifecycle.js";
-import { isAcceptedAgentDedupePayload, readGatewayDedupeEntry } from "./agent-dedupe.js";
+import { replayAgentTurnIfCached } from "./agent-dedupe.js";
 import { resolveAgentDeliveryPhase } from "./agent-delivery-phase.js";
 import type { RestoredCronContinuation } from "./agent-handler-helpers.js";
 import { waitForAgentJob } from "./agent-job.js";
@@ -43,46 +42,6 @@ type AgentTurnStartRequest = {
   io: AgentTurnIo;
   onRunObserved?: (runId: string) => void;
 };
-
-function replayAgentTurnIfCached(params: {
-  preflight: AgentRequestPreflight;
-  context: GatewayRequestHandlerOptions["context"];
-  io: AgentTurnIo;
-}): boolean {
-  const { agentDedupeKeys, runId } = params.preflight;
-  const cached = readGatewayDedupeEntry({
-    dedupe: params.context.dedupe,
-    keys: agentDedupeKeys,
-  });
-  if (!cached) {
-    return false;
-  }
-  if (cached.ok && isAcceptedAgentDedupePayload(cached.payload)) {
-    const cachedRunId = normalizeOptionalString(cached.payload.runId) ?? runId;
-    const cachedSessionKey = normalizeOptionalString(cached.payload.sessionKey);
-    const cachedAgentId = normalizeOptionalString(cached.payload.agentId);
-    const cachedRuntime = asOptionalRecord(cached.payload.runtime);
-    const admissionPending = typeof cached.payload.reservationId === "string";
-    params.io.emitAcceptance(
-      [
-        true,
-        {
-          runId: cachedRunId,
-          status: "in_flight" as const,
-          ...(cachedSessionKey ? { sessionKey: cachedSessionKey } : {}),
-          ...(cachedAgentId ? { agentId: cachedAgentId } : {}),
-          ...(cachedRuntime ? { runtime: cachedRuntime } : {}),
-          ...(admissionPending ? { admissionPending: true } : {}),
-        },
-        undefined,
-      ],
-      { cached: true, runId: cachedRunId },
-    );
-  } else {
-    params.io.emitAcceptance([cached.ok, cached.payload, cached.error], { cached: true });
-  }
-  return true;
-}
 
 export function createAgentTurnService(
   { context, isWebchatConnect }: Pick<GatewayRequestHandlerOptions, "context" | "isWebchatConnect">,
